+++ src/table/pagination-table/components/__tests__/TableWrapper.spec.tsx
@@ -4,7 +4,7 @@ import { stardust } from '@nebula.js/stardust';
 import TableWrapper from '../TableWrapper';
 import TableBodyWrapper from '../body/TableBodyWrapper';
 import TableHeadWrapper from '../head/TableHeadWrapper';
-import * as handleKeyPress from '../../../utils/handle-key-press';
+import * as handleKeyPress from '../../../utils/handle-keyboard';
 import * as handleScroll from '../../../utils/handle-scroll';
 import { TableLayout, TableData, PageInfo, SetPageInfo, Announce, Column } from '../../../../types';
 import TestWithProviders from '../../../../__test__/test-with-providers';
